Use Local instead of Handle
To make node-spotify compile with NodeJS 0.10 a lot of v8::Handles have
to be explicit v8::Locals.

I'm not super sure if this is correct this way (usually I try to use the "lowest"
interface possible), but since a lot of Nan functions are built like this I guess it's ok.