Bug #101 Fixes NodeJS audio playback
The V8 keys to set some properties on the buffer object were marked as
`static` which does not work as intended. Now they are created for each call.

Future implementations should probably add a new wrapped class for the buffer
object or cache the keys in `Persistent` handles instead of recreating them all
the time (puts pressure on the GC).